Adjust rustfmt setup
Make sure we are using the same rustfmt configuration as would be used
for formatting the host file itself.

On stable, we do that by setting the current directory to the directory
of the host file.

On nightly, we in addition capture rustfmt configuration, write it to
the temp file and turn on `normalize_doc_attributes`.