🔖Allow generating plain comments via the identifier trick
Adding a struct that can be `quote!`-ed to produce a plain Rust comment.
Internally, it generates a magic identifier and a string literal after
it. When we render token stream, we match on this identifier and
append a plain Rust comment to the output stream (and then let the
`rustfmt` to do the formatting in the end).

Closes #3